requirements.txt
PyQt5>=5.15.0
qt-material>=2.12
Pillow>=8.0.0
PyPDF2>=2.0.0
python-docx>=0.8.11
openpyxl>=3.0.0
pandas>=1.3.0
xlsxwriter>=3.0.0 